# src/main/java/org/springframework/petmanagement/rest/controller/ClinicController.java
package org.springframework.petmanagement.rest.controller;

import java.util.Optional;
import java.util.UUID;

import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.petmanagement.mapper.ClinicMapper;
import org.springframework.petmanagement.model.Clinic;
import org.springframework.petmanagement.rest.api.ClinicsApi;
import org.springframework.petmanagement.rest.dto.ClinicDto;
import org.springframework.petmanagement.rest.dto.ClinicFieldsDto;
import org.springframework.petmanagement.rest.dto.ClinicPageDto;
import org.springframework.petmanagement.service.ClinicService;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.util.UriComponentsBuilder;

@RestController
@CrossOrigin(exposedHeaders = "errors, content-type")
public class ClinicController implements ClinicsApi {

    private final ClinicService clinicService;
    private final ClinicMapper clinicMapper;

    public ClinicController(ClinicService clinicService, ClinicMapper clinicMapper) {
        this.clinicService = clinicService;
        this.clinicMapper = clinicMapper;
    }

    @PreAuthorize("hasAnyRole(@roles.ADMIN, @roles.CLINIC_ADMIN)")
    @Override
    public ResponseEntity<ClinicPageDto> listClinics(Integer page, Integer size) {
        Pageable pageable = PageRequest.of(page, size);
        Page<Clinic> clinics = clinicService.listClinics(pageable);
        return ResponseEntity.ok(clinicMapper.toClinicPageDto(clinics));
    }

    @PreAuthorize("hasAnyRole(@roles.ADMIN, @roles.CLINIC_ADMIN)")
    @Override
    public ResponseEntity<ClinicDto> getClinic(UUID clinicId) {
        Optional<Clinic> clinicOpt = clinicService.getClinic(clinicId);
        return clinicOpt
            .map(clinic -> ResponseEntity.ok(clinicMapper.toClinicDto(clinic)))
            .orElseGet(() -> ResponseEntity.notFound().build());
    }

    @PreAuthorize("hasRole(@roles.ADMIN)")
    @Override
    public ResponseEntity<ClinicDto> addClinic(ClinicFieldsDto clinicFieldsDto) {
        Clinic saved = clinicService.createClinic(clinicFieldsDto);

        HttpHeaders headers = new HttpHeaders();
        headers.setLocation(
            UriComponentsBuilder.newInstance()
                .path("/clinics/{id}")
                .buildAndExpand(saved.getId())
                .toUri()
        );
        return new ResponseEntity<>(clinicMapper.toClinicDto(saved), headers, HttpStatus.CREATED);
    }

    @PreAuthorize("hasRole(@roles.ADMIN)")
    @Override
    public ResponseEntity<ClinicDto> updateClinic(UUID clinicId, ClinicFieldsDto clinicFieldsDto) {
        try {
            Clinic updated = clinicService.updateClinic(clinicId, clinicFieldsDto);
            return ResponseEntity.ok(clinicMapper.toClinicDto(updated));
        } catch (IllegalArgumentException ex) {
            return ResponseEntity.notFound().build();
        }
    }

    @PreAuthorize("hasRole(@roles.ADMIN)")
    @Override
    public ResponseEntity<Void> deleteClinic(UUID clinicId) {
        try {
            clinicService.deleteClinic(clinicId);
            return ResponseEntity.noContent().build();
        } catch (IllegalArgumentException ex) {
            return ResponseEntity.notFound().build();
        }
    }
}