# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Pet Management API
  description: Pet Management API with health logs, prescriptions, item master, and JWT authentication.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: '2'
servers:
  - url: http://localhost:9966/petmanagement/api/v2
...
security:
  - bearerAuth: []
paths:
  /auth/signin:
    post:
      tags: [auth]
      operationId: authenticateUser
      summary: Sign in with Email and Password
      description: Returns JWT access token and refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/signup:
    post:
      tags: [auth]
      operationId: registerUser
      summary: Register a new user (Public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        201:
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        409:
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Get current authenticated user details
      description: Used to check if the current token is valid (Login Check).
      responses:
        200:
          description: OK (Token is valid)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        401:
          description: Unauthorized (Token is invalid or expired)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /auth/refreshtoken:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh Access Token
      description: Generates a new access token using a valid refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        200:
          description: Token Refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        403:
          description: Refresh token is invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /pets:
    get:
      tags: [pet]
      operationId: listPets
      summary: Lists all pets with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetPage'
    post:
      tags: [pet]
      operationId: addPet
      summary: Add a new pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
  /pets/{petId}:
    get:
      tags: [pet]
      operationId: getPet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags: [pet]
      operationId: updatePet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetFields'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
    delete:
      tags: [pet]
      operationId: deletePet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
  /clinics:
    get:
      tags: [clinic]
      operationId: listClinics
      summary: List all clinics with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicPage'
    post:
      tags: [clinic]
      operationId: addClinic
      summary: Add a new clinic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinic'
  /clinics/{clinicId}:
    get:
      tags: [clinic]
      operationId: getClinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinic'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags: [clinic]
      operationId: updateClinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClinicFields'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinic'
    delete:
      tags: [clinic]
      operationId: deleteClinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
...
  /items:
    get:
      tags: [item]
      operationId: listItems
      summary: List item master data with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemPage'
    post:
      tags: [item]
      operationId: addItem
      summary: Add new item master data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index (0..N)
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: The size of the page to be returned
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
  schemas:
...
    SigninRequest:
          type: object
          properties:
            email:
              type: string
              format: email
              maxLength: 255
              example: owner@example.com
            password:
              type: string
              format: password
              example: password123
          required:
            - email
            - password
    JwtResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        refreshToken:
          type: string
        expiresIn:
          type: integer
        userId:
          type: string
          format: uuid
      required:
        - accessToken
        - tokenType
        - refreshToken
        - userId
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).+$'
      required:
        - email
        - password
    SignupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
      required:
        - id
        - username
        - email
...
    TokenRefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    TokenRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
      required:
        - accessToken
        - refreshToken
        - tokenType

    # Page
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          description: The size of the page
        totalElements:
          type: integer
          description: Total number of elements across all pages
        totalPages:
          type: integer
          description: Total number of pages
        number:
          type: integer
          description: Current page number (0-based)
      required:
        - size
        - totalElements
        - totalPages
        - number
...
    PetPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Pet'
          required:
            - content
    ClinicPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Clinic'
          required:
            - content
...
    ItemPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Item'
          required:
            - content
...
    ProblemDetail:
      title: Problem Detail
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
    RoleName:
      type: string
      enum: [owner, vet, nurse, receptionist, clinic_admin, admin]
    PetType:
      type: string
      enum: [dog, cat, rabbit, hamster, bird, turtle, fish]
    PetSex:
      type: string
      enum: [male, female, unknown]
...
    ItemCategory:
      type: string
      enum: [food, treat, supplement, pad, toy, collar, hygiene, medical, other]
    Role:
      title: Role
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          $ref: '#/components/schemas/RoleName'
      required:
        - id
        - name
    ClinicFields:
      title: Clinic Fields
      type: object
      properties:
        name:
          type: string
          example: 'Sunflower Vet Clinic'
        telephone:
          type: string
          example: '03-1234-5678'
        address:
          type: string
          example: 'Tokyo, Shibuya 1-1'
        websiteUrl:
          type: string
          example: 'https://example.com'
        openingHours:
          type: string
          example: '9:00 - 18:00'
        note:
          type: string
      required:
        - name
    Clinic:
      allOf:
        - $ref: '#/components/schemas/ClinicFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - createdAt
            - updatedAt
...
    ItemFields:
      title: Item Fields
      type: object
      properties:
        name:
          type: string
          example: 'Premium Dog Food'
        category:
          $ref: '#/components/schemas/ItemCategory'
        note:
          type: string
        metadata:
          type: object
          additionalProperties: true
          description: JSONB metadata for external links or details
      required:
        - name
        - category
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true 
            createdAt: 
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - createdAt
            - updatedAt
...
    PetFields:
      title: Pet fields
      type: object
      properties:
        name:
          type: string
          maxLength: 30
          example: ポチ
        birthDate:
          type: string
          format: date
          example: '2020-01-01'
        sex:
          $ref: '#/components/schemas/PetSex'
        type:
          $ref: '#/components/schemas/PetType'
        userId:
          type: string
          format: uuid
          description: The owner ID
      required:
        - name
        - type
        - userId
    Pet:
      allOf:
        - $ref: '#/components/schemas/PetFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - createdAt
            - updatedAt
    UserBase:
          type: object
          properties:
            username:
              type: string
              minLength: 1
              maxLength: 20
              example: john.doe
            firstName:
              type: string
              maxLength: 50
            lastName:
              type: string
              maxLength: 50
            firstNameKana:
              type: string
              maxLength: 50
              pattern: '^[ァ-ヶー]*$'
            lastNameKana:
              type: string
              maxLength: 50
              pattern: '^[ァ-ヶー]*$'
            email:
              type: string
              format: email
              maxLength: 255
            postalCode:
              type: string
              maxLength: 8
              pattern: '^$|^\d{3}-?\d{4}$'
            prefecture:
              type: string
              maxLength: 10
            city:
              type: string
              maxLength: 80
            address:
              type: string
              maxLength: 255
            telephone:
              type: string
              maxLength: 20
              pattern: '^$|^[0-9-]*$'
          required:
            - email
            - username
...
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            roles:
              type: array
              items:
                $ref: '#/components/schemas/Role'
            enabled:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - id
            - roles
            - enabled
            - createdAt
            - updatedAt
...